% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\name{accuracy}
\alias{accuracy}
\title{extract (internal)}
\usage{
accuracy(embed, test_set, outcome)
}
\arguments{
\item{param1}{Description of the first parameter.}

\item{param2}{Description of the second parameter.}

\item{param3}{Description of the first parameter.}
}
\value{
What the function returns.
}
\description{
extract
}
\examples{
testfunction(1,2) # An example of how to use the function.
}
