% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{clean (internal)}
\usage{
clean(data, outcome, column, duplicates = "average")
}
\arguments{
\item{param1}{Description of the first parameter.}

\item{param2}{Description of the second parameter.}
}
\value{
What the function returns.
}
\description{
clean the given tibble such that it only contains the relevant columns)
}
\examples{
testfunction(1,2) # An example of how to use the function.
}
