% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{core_tensor}
\alias{core_tensor}
\title{clean (internal)}
\usage{
core_tensor(data, right_variables, proj_list, left_variable = NA)
}
\arguments{
\item{param1}{Description of the first parameter.}

\item{param2}{Description of the second parameter.}

\item{param3}{Description of the first parameter.}

\item{param4}{Description of the second parameter.}
}
\value{
What the function returns.
}
\description{
Description of what your function does. This part can span multiple lines
and will appear in the Details section of the documentation.
}
\examples{
testfunction(1,2) # An example of how to use the function.
}
