% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.R
\name{embed}
\alias{embed}
\title{embedding with a tensor}
\usage{
embed(interaction_model, k, max_iter = 100, tol = 0.001)
}
\arguments{
\item{param1}{Description of the first parameter.}

\item{param2}{Description of the second parameter.}

\item{param3}{Description of the first parameter.}

\item{param4}{Description of the second parameter.}

\item{param5}{Description of the first parameter.}

\item{param6}{Description of the second parameter.}
}
\value{
What the function returns.
}
\description{
Description of what your function does. This part can span multiple lines
and will appear in the Details section of the documentation.
}
\examples{
testfunction(1,2) # An example of how to use the function.
}
