% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{extract}
\alias{extract}
\title{extract (internal)}
\usage{
extract(new_data, proj, column)
}
\arguments{
\item{param1}{Description of the first parameter.}

\item{param2}{Description of the second parameter.}

\item{param3}{Description of the first parameter.}
}
\value{
What the function returns.
}
\description{
extract
}
\examples{
testfunction(1,2) # An example of how to use the function.
}
