% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetic.R
\name{matrix_to_tensor}
\alias{matrix_to_tensor}
\title{Transfer a matrix to a tensor (internal)}
\usage{
matrix_to_tensor(matrix, name)
}
\arguments{
\item{param1}{Description of the first parameter.}

\item{param2}{Description of the second parameter.}
}
\value{
What the function returns.
}
\description{
Description of what your function does. This part can span multiple lines
and will appear in the Details section of the documentation.
}
\examples{
testfunction(1,2) # An example of how to use the function.
}
