% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svds.R
\name{tensor_svds}
\alias{tensor_svds}
\title{tensor-vector multiplication}
\usage{
tensor_svds(edges, mode, edge_value = NA, irrelevant = NA, dim = c(1, 1, 1))
}
\arguments{
\item{param1}{Description of the first parameter.}

\item{param2}{Description of the second parameter.}

\item{param3}{Description of the first parameter.}

\item{param4}{Description of the second parameter.}

\item{param5}{Description of the second parameter.}
}
\value{
What the function returns.
}
\description{
Description of what your function does. This part can span multiple lines
and will appear in the Details section of the documentation.
}
\examples{
testfunction(1,2) # An example of how to use the function.
}
